requirements.txt
anyio==4.4.0
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cheroot==10.0.1
click==8.1.7
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
Flask==3.0.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipython==8.12.3
itsdangerous==2.2.0
jaraco.functools==4.0.2
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==10.4.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
outcome==1.3.0.post0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pipenv==2024.0.1
pipreqs==0.5.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-telegram-bot==21.4
pyzmq==26.1.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
selenium==4.23.1
setuptools==72.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
virtualenv==20.26.3
wcwidth==0.2.13
web.py==0.62
webbot==0.34
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
yarg==0.1.9
